requirements.txt
rasterio
numpy
pandas
requests
tqdm
pillow
torch
torchvision
